requirements.txt
pyTelegramBotAPI
Pillow
schedule
flask